Clean up table view cell UI.
