+++ Practical Example/Neil Hiddink/README.md
@@ -1,4 +1,3 @@
-
 # Practical Interview Project Submission
 ## Neil Hiddink
