--- Practical Example/Neil Hiddink/RWArticles/RWContent/Models/RWContentTableViewCell.swift
@@ -1,17 +1,0 @@
-//
-//  ContentTableViewCell.swift
-//  RWContent
-//
-//  Created by Neil Hiddink on 8/17/20.
-//  Copyright © 2020 Neil Hiddink. All rights reserved.
-//
-
-import UIKit
-
-class RWContentTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var detailLabel: UILabel!
-    @IBOutlet weak var previewImage: UIImageView!
-    
-}
