+++ Practical Example/Neil Hiddink/RWContent/RWContent/Models/RWClient.swift
