+++ Practical Example/Neil Hiddink/RWContent/RWContent/Models/RWResponse.swift
