+++ Practical Example/Neil Hiddink/RWContent/RWContent/Models/RWVideo.swift
